Split up history file
